more advanced canHandle
Video handlers can now use the DOM node as well as player URL to check whether
they can handle given Flash instance.

The code for getting flashvars in jwplayer.js is simplified, and no longer
fails when flashvars is present in neither OBJECT or EMBED/PARAM tag. This bug
was not discovered before because it previously filtered based on Player URL,
and therefore avoided the problem on the sites I tested.