increase specificity of CSS selector
A particular website uses a CSS rule "a:link {background: transparent}",
and since this has higher specificity than a class selector (a pseudoclass
has equal value as a class, and this one has an element in addition),
the video control background images are not displayed. Granted that this
website is to blame, but it should be a good idea to use a high specifity
selector for something special like the buttons.